+++ src/Spydersoft.Platform.Hosting/Spydersoft.Platform.Hosting/StartupExtensions/TelemetryExtensions.cs
@@ -1,5 +1,4 @@
-﻿using Microsoft.AspNetCore.Authentication.JwtBearer;
-using Microsoft.AspNetCore.Builder;
+﻿using Microsoft.AspNetCore.Builder;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
@@ -15,9 +14,9 @@ using Spydersoft.Platform.Hosting.Options;
 using System.Diagnostics.Metrics;
 using System.Reflection;
 
-namespace Spydersoft.Platform.Hosting;
+namespace Spydersoft.Platform.Hosting.StartupExtensions;
 
-public static class StartupExtensions
+public static class TelemetryExtensions
 {
     #region Public Startup Extensions
     public static void AddSpydersoftSerilog(this WebApplicationBuilder appBuilder)
@@ -50,67 +49,11 @@ public static class StartupExtensions
             .WithLogging(builder => ConfigureLogging(builder, telemetryOptions));
     }
 
-    public static bool AddSpydersoftIdentity(this WebApplicationBuilder appBuilder)
-    {
-        var authInstalled = false;
-        var identityOption = new IdentityOptions();
-        appBuilder.Configuration.GetSection(IdentityOptions.SectionName).Bind(identityOption);
-
-        if (identityOption.Enabled && identityOption.Authority != null)
-        {
-            appBuilder.Services
-                .AddAuthentication(o =>
-                {
-                    o.DefaultScheme = JwtBearerDefaults.AuthenticationScheme;
-                })
-                .AddJwtBearer(o =>
-                {
-                    o.IncludeErrorDetails = true;
-
-                    o.Authority = identityOption.Authority;
-                    o.TokenValidationParameters = new Microsoft.IdentityModel.Tokens.TokenValidationParameters
-                    {
-                        ValidAudiences =
-                                        [
-                                            identityOption.ApplicationName
-                                        ],
-                        ValidIssuers =
-                                        [
-                                            identityOption.Authority
-                                        ]
-                    };
-                });
-
-            appBuilder.Services.AddAuthorization();
-            authInstalled = true;
-        }
-        return authInstalled;
-    }
-
-    public static IApplicationBuilder UseAuthentication(this IApplicationBuilder app, bool authInstalled)
-    {
-        if (authInstalled)
-        {
-            app.UseAuthentication();
-        }
-        return app;
-    }
-
-    public static IApplicationBuilder UseAuthorization(this IApplicationBuilder app, bool authInstalled)
-    {
-        if (authInstalled)
-        {
-            app.UseAuthorization();
-        }
-        return app;
-    }
-
-
     #endregion
 
 
     #region Private Startup Helpers
-    private static void ConfigureTracing(TracerProviderBuilder builder, IConfiguration configuration, TelemetryOptions options)
+    private static void ConfigureTracing(TracerProviderBuilder builder, ConfigurationManager configuration, TelemetryOptions options)
     {
 
         // Ensure the TracerProvider subscribes to any custom ActivitySources.