+++ src/Spydersoft.Platform.Hosting/docs/CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* Standardized HealthCheck options
+* Test Project (Spydersoft.Platform.Hosting.ApiTests)
+
 ### Changed
 
 ### Removed